LayoutTests/fast/scrolling/scroll-anchoring-in-overflow-scroll.html

<!DOCTYPE html><!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true ] --> 
<html>
<head>
    <style>
        .scroller {
            width: 300px;
            height: 400px;
            border: 1px solid black;
            overflow-y: scroll;
        }
        
        .changing {
            height: 300px;
            background-color: orange;
        }

        body.changed .changing {
            height: 400px;
        }

        .anchor {
            height: 300px;
            background-color: limegreen;
        }
        
        .spacer {
            height: 500px;
        }
        
    </style>
    <script src="../../resources/ui-helper.js"></script>
    <script src="../../resources/js-test.js"></script>
    <script>
        jsTestIsAsync = true;

        let scroller;
        window.addEventListener('load', async () => {
            scroller = document.querySelector('.scroller');
            scroller.scrollTo(0, 310);
            // should anchor to 'anchor'

            shouldBe('scroller.scrollTop', '310');
            document.body.classList.toggle('changed');

            await UIHelper.renderingUpdate();

            shouldBe('scroller.scrollTop', '410');

            document.body.classList.toggle('changed');

            await UIHelper.ensurePresentationUpdate();

            shouldBe('scroller.scrollTop', '310');

            finishJSTest();
        }, false);
    </script>
</head>
<body>
    <div class="scroller">
        <div class="changing"></div>
        <div class="anchor"></div>
        <div class="spacer"></div>
    </div>
<div id="console"></div>
</body>
</html>
